Create Persistdir If It Doesn't Exist
Just run mkdirs if the persist dir doesn't exist.  I decided
to do this in Batman but I could have decided that this should
only happy in Puppet or something.  I decided to do it in Batman
because Batman's whole point is to ease configuation, and if you have
to do a bunch of work to get the world into a state that you can
even *run* batman, it's not doing it's job.

Closes #7